+++ tomcat/src/main/java/org/apache/catalina/SessionManager.java
@@ -1,8 +1,9 @@
-package org.apache.coyote.util;
+package org.apache.catalina;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
+import org.apache.catalina.util.Cookie;
 
 public class SessionManager {
