+++ tomcat/src/main/java/org/apache/catalina/util/StaticResourcePathGenerator.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.util;
+package org.apache.catalina.util;
 
 public class StaticResourcePathGenerator {
