+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,18 +1,13 @@
 package org.apache.coyote.http11;
 
-import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.exception.UncheckedServletException;
-import com.techcourse.model.User;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.Socket;
-import java.util.Optional;
 import org.apache.coyote.Processor;
-import org.apache.coyote.util.Cookie;
-import org.apache.coyote.util.Session;
-import org.apache.coyote.util.SessionManager;
-import org.apache.coyote.util.StaticResourcePathGenerator;
+import org.apache.catalina.core.ApplicationProcessor;
+import org.apache.catalina.util.StaticResourcePathGenerator;
 import org.apache.coyote.util.request.HttpRequest;
 import org.apache.coyote.util.request.HttpRequestParser;
 import org.apache.coyote.util.response.HttpContentTypeResolver;
@@ -20,6 +15,10 @@ import org.apache.coyote.util.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * Coyote 영역의 HTTP/1.1 프로토콜 처리기
+ * 네트워크 I/O와 HTTP 프로토콜 처리에만 집중
+ */
 public class Http11Processor implements Runnable, Processor {
 
     private static final Logger log = LoggerFactory.getLogger(Http11Processor.class);
@@ -46,17 +45,36 @@ public class Http11Processor implements Runnable, Processor {
                 respond(HttpResponse.of("HTTP/1.1 404 Not Found", "static/404.html"), outputStream);
                 return;
             }
+            
+            // API 요청은 Catalina의 ApplicationProcessor로 위임
             if (handleApiRequest(request, outputStream)) {
                 return;
             }
+            
+            // 정적 리소스 처리
             String resourcePath = StaticResourcePathGenerator.generate(request.getPath());
             if (handleStaticResourceRequest(resourcePath, outputStream)) {
                 return;
             }
+            
             respond(HttpResponse.of("HTTP/1.1 404 Not Found", "static/404.html"), outputStream);
         } catch (IOException | UncheckedServletException e) {
             log.error(e.getMessage(), e);
         }
+    }
+
+    private boolean handleApiRequest(HttpRequest request, OutputStream outputStream) throws IOException {
+        if ("/login".equals(request.getPath())) {
+            HttpResponse loginResponse = ApplicationProcessor.processLogin(request);
+            respond(loginResponse, outputStream);
+            return true;
+        }
+        if ("/register".equals(request.getPath()) && "POST".equals(request.getMethod())) {
+            HttpResponse registerResponse = ApplicationProcessor.processRegister(request);
+            respond(registerResponse, outputStream);
+            return true;
+        }
+        return false;
     }
 
     private boolean handleStaticResourceRequest(String resourcePath, OutputStream outputStream) throws IOException {
@@ -74,68 +92,6 @@ public class Http11Processor implements Runnable, Processor {
         return false;
     }
 
-    private boolean handleApiRequest(HttpRequest request, OutputStream outputStream) throws IOException {
-        if ("/login".equals(request.getPath())) {
-            if ("GET".equals(request.getMethod())) {
-                Session existingSession = request.getSession(false);
-                if (existingSession != null && getUser(existingSession) != null) {
-                    respond(HttpResponse.redirect("/index.html"), outputStream);
-                    return true;
-                }
-                respond(HttpResponse.of("HTTP/1.1 200 OK", "static/login.html"), outputStream);
-                return true;
-            }
-            if ("POST".equals(request.getMethod())) {
-                HttpResponse loginResponse = processLoginMemberInfo(request);
-                respondWithSession(loginResponse, outputStream, request);
-                return true;
-            }
-        }
-        if ("/register".equals(request.getPath()) && "POST".equals(request.getMethod())) {
-            HttpResponse registerResponse = processRegisterMember(request);
-            respondWithSession(registerResponse, outputStream, request);
-            return true;
-        }
-        return false;
-    }
-
-    private HttpResponse processLoginMemberInfo(HttpRequest httpRequest) {
-        String account = httpRequest.getQueryValue("account")
-                .orElse(null);
-        String password = httpRequest.getQueryValue("password")
-                .orElse(null);
-        if (account == null || password == null) {
-            return HttpResponse.redirect("401.html");
-        }
-        Optional<User> userOpt = InMemoryUserRepository.findByAccount(account);
-        if (userOpt.isEmpty() || !userOpt.get().checkPassword(password)) {
-            return HttpResponse.redirect("401.html");
-        }
-        User user = userOpt.get();
-        final Session session = httpRequest.getSession(true);
-        session.setAttribute("user", user);
-        HttpResponse response = HttpResponse.redirect("/index.html");
-        response.addCookie(SessionManager.JSESSIONID, session.getId());
-        log.info("User: {}", user);
-        return response;
-    }
-
-    private HttpResponse processRegisterMember(HttpRequest request) {
-        String account = request.getQueryValue("account")
-                .orElse(null);
-        String email = request.getQueryValue("email")
-                .orElse(null);
-        String password = request.getQueryValue("password")
-                .orElse(null);
-        if (account == null || email == null || password == null) {
-            return HttpResponse.redirect("static/404.html");
-        }
-        User user = new User(account, password, email);
-        InMemoryUserRepository.save(user);
-        log.info("회원가입 성공! 아이디: {}", user.getAccount());
-        return HttpResponse.redirect("/index.html");
-    }
-
     private byte[] readPathFile(String requestPath) {
         try (InputStream inputStream = getClass().getClassLoader().getResourceAsStream(requestPath)) {
             if (inputStream == null) {
@@ -151,19 +107,5 @@ public class Http11Processor implements Runnable, Processor {
         outputStream.write(httpResponse.createHeader().getBytes());
         outputStream.write(httpResponse.getBody());
         outputStream.flush();
-    }
-
-    private void respondWithSession(HttpResponse httpResponse, OutputStream outputStream, HttpRequest request)
-            throws IOException {
-        Cookie cookie = request.getCookie();
-        if (!SessionManager.hasValidSessionId(cookie)) {
-            String sessionId = SessionManager.generateSessionId();
-            httpResponse.addCookie(SessionManager.JSESSIONID, sessionId);
-        }
-        respond(httpResponse, outputStream);
-    }
-
-    private User getUser(Session session) {
-        return (User) session.getAttribute("user");
     }
 }
