+++ tomcat/src/main/java/org/apache/coyote/util/request/HttpRequestParser.java
@@ -8,7 +8,7 @@ import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.coyote.util.Cookie;
+import org.apache.catalina.util.Cookie;
 
 public class HttpRequestParser {
